requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
googlesearch_python==1.2.3
lxml==4.9.2
matplotlib==3.7.1
nltk==3.8.1
wordcloud==1.9.2
